PBD-2183 Deprecate existing implementation of AwsCodeBuildStep
This shouldn't really have been implemented as a `Configure`, as it means it doesn't play nicely with other `Step` classes in terms of ordering.

There is also a problem with the triggering failing if a sleep isn't included.

In future this might be re-implemented as a `Step`, but for now the advice is to manually define it in DSL as follows:

```
steps {awsCodeBuild projectName: 'foo-codebuild-project', ... }
```