PBD-4489 Upgrade Jenkins core and plugin versions
Unfortunately we've had to temporarily suspend testing of a couple of small bits of functionality, namely GitHub pull request triggers and build description setters. This is because the plugins they depend on haven't been updated for many years, and installing them in the test stack causes other plugins to be downgraded or uninstallable, breaking other tests. We will review this in future.